Add useResetModal
